requirements.txt
streamlit
pandas
Pillow
groq
python-dotenv